Restructure src/ according to transforms.
