+++ bin/stopify
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-node --harmony_tailcalls "`npm root -g`/Stopify/built/src/stopifyStandalone.js" $@
+node --harmony_tailcalls "`npm root -g`/Stopify/built/src/stopify.js" $@
