+++ src/common/markKnownFunctions.ts
