+++ src/cps/desugarLoop.ts
@@ -14,7 +14,7 @@
 
 import {NodePath, VisitNode, Visitor} from 'babel-traverse';
 import * as t from 'babel-types';
-import * as h from './helpers';
+import * as h from '../common/helpers';
 
 // Object containing the visitor functions
 const loopVisitor : Visitor = {
