+++ src/cps/desugarSwitch.ts
@@ -1,6 +1,6 @@
 import {NodePath, VisitNode, Visitor} from 'babel-traverse';
 import * as t from 'babel-types';
-import {letExpression} from './helpers';
+import {letExpression} from '../common/helpers';
 
 function ifTest(e: t.Expression,
   discriminant: t.Identifier,
