+++ src/interfaces/steppifyInterface.ts
