+++ src/interfaces/stopifyInterface.ts
