+++ src/yield/renameConstructor.ts
